\name{LocallyStationaryModels-package}
\alias{LocallyStationaryModels-package}
\alias{LocallyStationaryModels}
\docType{package}
\title{
\packageTitle{LocallyStationaryModels}
}
\description{
\packageDescription{LocallyStationaryModels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LocallyStationaryModels}
\packageIndices{LocallyStationaryModels}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LocallyStationaryModels}

Maintainer: \packageMaintainer{LocallyStationaryModels}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
