% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plot.parameters}
\alias{plot.parameters}
\title{Plot parameters}
\usage{
\method{plot}{parameters}(allpoints)
}
\arguments{
\item{allpoints}{a dataframe containing the coordinates and the value of the parameters}
}
\description{
sequential plot of all the paramters obtained by smoothing
}
\details{
given a dataframe containing the coordinates and the the value of the parameters for a set of points
this functions provides the plot in the plane of the trend of all the parameters. You can use this function, for example, with
the dataframe returned by plot.lsm to visualize other parameters from the four that are already plot by the previous function
}
\examples{
plot.parameters(mypoints)
}
